emref to __init__.py
